surreal: Create Datasets with Hidden Images in Residual Plots

Implements the "Residual (Sur)Realism" algorithm described by Stefanski (2007) doi:10.1198/000313007X190079 to generate datasets that reveal hidden images or messages in their residual plots. It offers both predefined datasets and tools to embed custom text or images into residual structures. Allowing users to create intriguing visual demonstrations for teaching model diagnostics.

See also

Author

Maintainer: James Joseph Balamuta james.balamuta@gmail.com (ORCID) [copyright holder]